remove validation.py
