+++ pyblazing/pyblazing/apiv2/context.py
@@ -46,6 +46,8 @@ from enum import IntEnum
 
 import platform
 
+import sys
+
 jpype.addClassPath(
     os.path.join(os.getenv("CONDA_PREFIX"), "lib/blazingsql-algebra.jar")
 )
@@ -847,6 +849,41 @@ def convert_friendly_dtype_to_string(list_types):
     return list_types
 
 
+def kwargs_validation(kwargs, bc_api_str):
+    """
+    Validation of kwargs params when a bc API is called
+    """
+    # csv, parquet, orc, json params
+    if bc_api_str == "create_table":
+        full_kwargs = ['file_format', 'names', 'dtype', 'delimiter', 'skiprows', 'skipfooter',
+                       'lineterminator', 'header', 'nrows', 'skip_blank_lines', 'decimal',
+                       'true_values', 'false_values', 'na_values', 'keep_default_na',
+                       'na_filter', 'quotechar', 'quoting', 'doublequote', 'comment',
+                       'delim_whitespace', 'skipinitialspace', 'use_cols_indexes',
+                       'use_cols_names', 'byte_range_offset', 'byte_range_size',
+                       'compression', 'lines', 'stripes', 'skiprows', 'num_rows', 'use_index']
+        params_info = "https://docs.blazingdb.com/docs/create_table"
+
+    elif bc_api_str == "s3":
+        full_kwargs = ['name', 'bucket_name', 'access_key_id', 'secret_key', 'encryption_type',
+                       'session_token', 'root', 'kms_key_amazon_resource_name', 'endpoint_override', 'region']
+        params_info = "https://docs.blazingdb.com/docs/s3"
+
+    elif bc_api_str == "hdfs":
+        full_kwargs = ['name', 'host', 'port', 'user', 'kerb_ticket']
+        params_info = "https://docs.blazingdb.com/docs/hdfs"
+
+    elif bc_api_str == "gs":
+        full_kwargs = ['name', 'project_id', 'bucket_name', 'use_default_adc_json_file', 'adc_json_file']
+        params_info = "https://docs.blazingdb.com/docs/google-storage"
+    
+    for arg_i in kwargs.keys():
+        if arg_i not in full_kwargs:
+            print("The parameter \'" + arg_i + "\' does not exists. Please make sure you are using the correct parameter:")
+            print("To get the correct parameters, check:  " + params_info)
+            sys.exit()
+
+
 class BlazingTable(object):
     def __init__(
         self,
@@ -1491,6 +1528,7 @@ class BlazingContext(object):
 
         Docs: https://docs.blazingdb.com/docs/hdfs
         """
+        kwargs_validation(kwargs, "hdfs")
         return self.fs.hdfs(self.dask_client, prefix, **kwargs)
 
     def s3(self, prefix, **kwargs):
@@ -1539,6 +1577,7 @@ class BlazingContext(object):
 
         Docs: https://docs.blazingdb.com/docs/s3
         """
+        kwargs_validation(kwargs, "s3")
         return self.fs.s3(self.dask_client, prefix, **kwargs)
 
     def gs(self, prefix, **kwargs):
@@ -1571,6 +1610,7 @@ class BlazingContext(object):
 
         Docs: https://docs.blazingdb.com/docs/google-storage
         """
+        kwargs_validation(kwargs, "gs")
         return self.fs.gs(self.dask_client, prefix, **kwargs)
 
     def show_filesystems(self):
